// src/app/api/model/partnership-opportunity-photo-dto-in.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PartnershipOpportunityPhotoDtoIn { 
    /**
     * Identifier of an EXISTING photo to keep (mutually exclusive with uploadId)
     */
    id?: number;
    /**
     * Indicates if this photo is the cover image
     */
    isCover: boolean;
    /**
     * Display order of the photo (0-based)
     */
    orderNumber: number;
    /**
     * Tracked upload id from POST /upload/signed-url for a NEW photo (mutually exclusive with id)
     */
    uploadId?: string;
}