// src/app/api/model/platform-dto.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContentTypeDto } from './content-type-dto';


export interface PlatformDto { 
    active: boolean;
    contentTypes: Set<ContentTypeDto>;
    id?: number;
    logoUrl?: string;
    name?: string;
}