// src/app/api/model/service-type-dto-out.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ServiceTypeDtoOut { 
    category?: string;
    description?: string;
    id?: number;
    name?: string;
    originalCategory?: string;
    originalDescription?: string;
    originalName?: string;
}